app/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
SQLAlchemy>=2.0.30
httpx>=0.27.0
python-dateutil>=2.9.0
pytz>=2024.1
cachetools>=5.3.3
numpy>=1.26.0
astronomy-engine>=2.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0
geopy==2.4.1
slowapi
timezonefinder